Exercise 2
Search and Write program. Write a program that would take in a text file and a search word as input. The program should write an output file with the line number and the start indices of the search term in that line in the format <line-no>:<start-index-1>,<start-index-2>. The name of the output file should be the <search-word>.locations.